rename repo ai helpers
